+++ lua/nvim-chezmoi/chezmoi/commands/encrypt.lua
@@ -1,0 +1,14 @@
+local command = require("nvim-chezmoi.chezmoi.command")
+
+---@class ChezmoiEncrypt: ChezmoiCommand
+local M = setmetatable({
+  cmd = "encrypt",
+}, {
+  __index = command,
+})
+
+function M:exec(stdin)
+  return command.exec(self, nil, stdin)
+end
+
+return M